sauce bottle button
This is my homework for Creative Computing Class. I want to create a button that looks not like a button. Later, I chose a sauce bottle to be the model of the button. There are two metal sheets inside the bottle. When I squeeze the bottle, the metal sheets will be connected and the circuit will be closed. Also when being squeezed, the bottle will make the sound of airflow, just like a button that can breathe. I also have an idea to continually modify the sauce bottle to be like a potentiometer which resistance will be different when it is squeezed by different pressure.